requirement.txt
aniso8601==9.0.1
apturl==0.5.2
asgiref==3.7.2
bcrypt==3.2.0
blinker==1.7.0
Brlapi==0.8.3
CacheControl==0.13.1
cachetools==5.3.2
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
duplicity==0.8.21
fasteners==0.14.1
filelock==3.13.1
firebase-admin==6.3.0
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
future==0.18.2
gcloud==0.18.3
google-api-core==2.15.0
google-api-python-client==2.113.0
google-auth==2.26.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
grpcio-status==1.60.0
gyp==0.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
jws==0.1.3
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.3
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.7
netifaces==0.11.0
numpy==1.26.3
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pipenv==2023.11.15
platformdirs==4.1.0
proto-plus==1.23.0
protobuf==4.25.1
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.20.1
pycrypto==2.6.1
pycryptodome==3.20.0
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.8.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
Pyrebase==3.0.27
Pyrebase4==4.7.1
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-jwt==2.0.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
SecretStorage==3.3.1
six==1.16.0
SQLAlchemy==2.0.25
sqlparse==0.4.4
sseclient==0.0.27
systemd-python==234
typing_extensions==4.9.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.25.0
wadllib==1.3.6
Werkzeug==3.0.1
xdg==5
xkit==0.0.0
zipp==1.0.0
